+++ flow.py
@@ -1,6 +1,7 @@
 import re
 import json
 from hat_manager import list_hats_by_team, add_memory_to_hat, search_memory
+from prompts import generate_openai_response, generate_openai_response_with_system
 from utils import generate_unique_hat_id
 
 import chainlit as cl
@@ -13,29 +14,34 @@ client = openai.OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
 
 
 async def run_team_flow(team_id, input_text):
+    #log the state of the current hat
+    current_hat = cl.user_session.get("current_hat")
+    cl.user_session.set("previous_hat", current_hat)
+    # Load and sort the team hats based on flow_order
     team_hats = list_hats_by_team(team_id)
+    team_hats = sorted(team_hats, key=lambda h: h.get("flow_order", 0))
+
     current_input = input_text
     conversation_log = []
     retry_counts = {}
 
-    i = 0
-    while i < len(team_hats):
-        hat = team_hats[i]
-        hat_name = hat.get('name')
-        hat_id = hat.get('hat_id')
-        qa_loop = hat.get('qa_loop', False)
-        retry_limit = hat.get('retry_limit', 0)
+    for hat in team_hats:
+        hat_name = hat.get("name", "Unnamed Hat")
+        hat_id = hat.get("hat_id")
+        qa_loop = hat.get("qa_loop", False)
+        retry_limit = hat.get("retry_limit", 0)
 
-
-
+        # --- Run the hat's response ---
         response_text = generate_openai_response(current_input, hat)
 
-        # 📝 Save memory with tags
+        # Save memory (input and output separately)
         add_memory_to_hat(hat_id, current_input, role="user")
         add_memory_to_hat(hat_id, response_text, role="bot")
 
-        await cl.Message(content=f"🧢 {hat_name} responded:\n{response_text}").send()
+        # Show the response in the chat
+        await cl.Message(content=f"🧢 **{hat_name}** responded:\n{response_text}").send()
 
+        # Log the conversation
         conversation_log.append({
             "hat_name": hat_name,
             "hat_id": hat_id,
@@ -43,62 +49,28 @@ async def run_team_flow(team_id, input_text):
             "output": response_text
         })
 
+        # Handle QA loop if enabled
         if qa_loop:
             handled = await handle_qa_loop(
                 hat, team_hats, conversation_log, retry_counts, retry_limit, team_id
             )
             if handled:
-                return  # Pause for user approval, QA loop handled
+                return  # Pause here if QA loop is triggered
 
+        # Pass the output to the next hat
         current_input = response_text
-        i += 1
 
-    log_text = "\n\n".join([
-        f"🧢 {entry['hat_name']}:\nInput: {entry['input']}\nOutput: {entry['output']}"
-        for entry in conversation_log
-    ])
-    await cl.Message(content=f"📜 **Full Conversation Log:**\n\n{log_text}").send()
-    await cl.Message(content="✅ Team flow complete!").send()
-
-
-def generate_openai_response(prompt: str, hat: dict):
-    hat_name = hat.get('name', 'Unnamed Agent')
-    hat_id = hat.get('hat_id')
-    instructions = hat.get('instructions', '')
-    role = hat.get('role', 'agent')
-    tools = ", ".join(hat.get('tools', [])) or "none"
-
-    memory_context = ""
-    relevant_memories = search_memory(hat_id, prompt, k=3)
-    if relevant_memories:
-        formatted = "\n".join([
-            f"{meta.get('role', 'unknown').capitalize()} ({meta.get('timestamp', 'no time')}): {doc}"
-            for doc, meta in relevant_memories if meta and doc
-        ])
-        memory_context = f"\n\nRelevant Memories:\n{formatted}"
-
-    system_prompt = f"""
-You are a {role} agent named '{hat_name}'.
-Your tools: {tools}.
-Instructions: {instructions}.
-{memory_context if memory_context else ''}
-""".strip()
-
-    response = client.chat.completions.create(
-        model=hat.get("model", "gpt-3.5-turbo"),
-        messages=[
-            {"role": "system", "content": system_prompt},
-            {"role": "user", "content": prompt}
-        ],
-        temperature=0.7,
-        max_tokens=1000
-    )
-
-    return response.choices[0].message.content
-
+    # # --- Final Summary ---
+    # log_text = "\n\n".join([
+    #     f"🧢 **{entry['hat_name']}**\n**Input:** {entry['input']}\n**Output:** {entry['output']}"
+    #     for entry in conversation_log
+    # ])
+    # await cl.Message(content=f"📜 **Full Team Conversation Log:**\n\n{log_text}").send()
+    await cl.Message(content="✅ **Team flow completed successfully!**").send()
 
 async def handle_qa_loop(hat, team_hats, conversation_log, retry_counts, retry_limit, team_id):
     response_text = conversation_log[-1]['output']
+
     if "#REVISION_REQUIRED" in response_text:
         retry_target_index = len(conversation_log) - 2
         retry_target = conversation_log[retry_target_index] if retry_target_index >= 0 else None
@@ -107,22 +79,36 @@ async def handle_qa_loop(hat, team_hats, conversation_log, retry_counts, retry_l
             retry_counts[retry_target['hat_id']] = retry_counts.get(retry_target['hat_id'], 0) + 1
 
             if retry_counts[retry_target['hat_id']] <= retry_limit:
-                await cl.Message(content=f"🔁 Critic requested revision. Retrying {retry_target['hat_name']}...").send()
+                await cl.Message(content=f"🔁 Critic requested revision. Retrying {retry_target['hat_name']} with feedback...").send()
 
                 prev_hat = next(h for h in team_hats if h['hat_id'] == retry_target['hat_id'])
-                retry_response = generate_openai_response(retry_target['input'], prev_hat)
+
+                # 🛠 Instead of resending original input blindly, attach Critic's feedback
+                improved_input = (
+                    f"{retry_target['input']}\n\n"
+                    f"Critic Feedback: {response_text}"
+                )
+
+                # Generate new response using improved input
+                retry_response = await generate_openai_response_with_system(
+                    user_prompt=retry_target['input'],
+                    system_prompt=f"Revision guidance: {response_text}",
+                    hat=prev_hat
+                )
+                
+                await cl.Message(content="🧠 **This was an improved attempt based on Critic feedback.**\n\nLet's see if it passes review this time!").send()
                 prev_hat_tags = prev_hat.get('memory_tags', [])
-                add_memory_to_hat(prev_hat['hat_id'], retry_target['input'], role="user", tags=prev_hat_tags, session=cl.user_session)
+                add_memory_to_hat(prev_hat['hat_id'], improved_input, role="user", tags=prev_hat_tags, session=cl.user_session)
                 add_memory_to_hat(prev_hat['hat_id'], retry_response, role="bot", tags=prev_hat_tags, session=cl.user_session)
                 
                 await cl.Message(content=f"🧢 {prev_hat['name']} retry responded:\n{retry_response}").send()
-                await cl.Message(content="🏷️ Want to tag this memory? Type: `tag last as <tag>`").send()
+
+                # Critic re-reviews the new retry
                 critic_response = generate_openai_response(retry_response, hat)
 
                 qa_tags = hat.get('memory_tags', [])
-
                 add_memory_to_hat(hat['hat_id'], retry_response, role="user", tags=qa_tags, session=cl.user_session)
-                add_memory_to_hat(hat['hat_id'], critic_response, role="bot" , tags=qa_tags, session=cl.user_session)
+                add_memory_to_hat(hat['hat_id'], critic_response, role="bot", tags=qa_tags, session=cl.user_session)
 
                 await cl.Message(content=f"🧢 {hat['name']} re-reviewed:\n{critic_response}").send()
 
@@ -135,9 +121,10 @@ async def handle_qa_loop(hat, team_hats, conversation_log, retry_counts, retry_l
                 cl.user_session.set("awaiting_user_approval", True)
                 cl.user_session.set("pending_critique_input", retry_response)
                 cl.user_session.set("pending_team_id", team_id)
-                return True  # Indicate we paused for approval
+                return True  # Pause for user approval
             else:
                 await cl.Message(content="⚠️ Retry limit reached. Proceeding.").send()
+
     elif "#APPROVED" in response_text:
         await cl.Message(content="✅ Critic approved!").send()
         await cl.Message(content="🧑‍⚖️ Approve or Retry? Type `approve` or `retry`.").send()
@@ -145,6 +132,7 @@ async def handle_qa_loop(hat, team_hats, conversation_log, retry_counts, retry_l
         cl.user_session.set("pending_critique_input", conversation_log[-1]['input'])
         cl.user_session.set("pending_team_id", team_id)
         return True
+
     else:
         await cl.Message(content="⚠️ No valid tag from Critic. Manual review needed.").send()
         await cl.Message(content="🧑‍⚖️ Approve or Retry? Type `approve` or `retry`.").send()
@@ -152,7 +140,9 @@ async def handle_qa_loop(hat, team_hats, conversation_log, retry_counts, retry_l
         cl.user_session.set("pending_critique_input", conversation_log[-1]['input'])
         cl.user_session.set("pending_team_id", team_id)
         return True
+
     return False
+
 
 
 def generate_team_from_goal(goal):
